Don't skip RDNs when reading from LDAP.  Just treat them like normal attributes.
Skipping RDNs was causing 'terraform apply' to always attempt to add the RDN attribute.
And when terraform attempted to add the RDN attribute, the ldap server would complain that the attribute already existed.

I didn't see a reason why RDNs should be skipped when reading from LDAP,
so I just removed the the logic to skip them.